several refactorings including tests
